+++ Aridos.BL/Models/DatosTemperatura.cs
@@ -11,10 +11,10 @@
         public int ID { get; set; }
         public int TipoSensor { get; set; }
         public string NumeroSerie { get; set; }
-        public double Temperatura { get; set; }
-        public double TAmbiente { get; set; }
-        public double TInterna { get; set; }
-        public double Bateria { get; set; }
+        public float Temperatura { get; set; }
+        public float TAmbiente { get; set; }
+        public float TInterna { get; set; }
+        public float Bateria { get; set; }
         public DateTime Fecha { get; set; }
     }
 }
